requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
pydantic>=2.9.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
httpx>=0.25.0
slowapi==0.1.9
google-genai>=1.68.0
requests==2.31.0
email-validator==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
boto3==1.34.162
reportlab
